Collision revamp, other tweaks.
The collision system now just sets a collider id, then we have a different
system that stops and starts movement based on whether the entity is colliding.

Also converted background stuff into entities so we can get rid of surface if
we like. Still am wanting better HiDPI handling in renderer or rendersys or
something...